[Update] Updated test cases for pages-code-of-conduct component from enzyme to rtl
